VS-13661: extract MaxCallback into its own dto file
MaxMessage.kt now holds only the message object graph (MaxUser,
MaxRecipient, MaxMessageBody, MaxMessage); the callback payload moves to
MaxCallback.kt so each dto file maps to one clear concern.
